Fix the crash bug.
While the master tries to activate instances of liveboxes
The slave is crashed without dead callback invocation (via AUL).
(or timing issue that the provider receives connection failed event first)

In this case,
activate_cb function of instance will be called with "result = NULL".
Then the activate_cb function will try to delete it. (just decrease the
refcnt)

But if the instance is referenced by the other functions,
(such as other requests resize, clicked, etc ...)
The refcnt will not reached to the ZERO even if the activate_cb was
called, so the instance will be kept in alive.

Now the problem is raised.
After executing above routines,
The data provider will forcely terminates the slave to recover this unexpected problem.
(because of the timing issue (or AUL missed send dead event).
--> Data provider just think that the current slave is not terminated but it
couldn't managed request correctly so it will terminates the problem
slave, to recover this problem, it will terminates the slave forcely.

Then the slave deactivate callback will be called and tries to delete
the instances which are recorded as a fault package,
because of the instance refcnt is not reached to ZERO, slave deactivate
callback will destroy thes instances again even if it is already cared
by previous routines.

After processing this routines again, the refcnt will be tained.
So who really references this instance will not be guaranted about this instance.

So, to fix this, I just separate the case of switch statement in the
slave_deactivated_cb.
I prevent decreasing when the state of an instance is "INST_DESTROYED".
(INST_DESTROYED means already someone cares this instance,
so don't try to destroy again from any other stages)

Change-Id: I7b41bcb69ecf8f0f0a82536ab9af57faf0d3b42a