Replace dbus with socket (connector)
And there is more minor patche is applied.
------------------------------------------
Keep the last state of the Buffer

When the slave is deactivated,
all instance on that slave will be destroyed.
And reactivating normal instances again.

While executing this procedure,
master will deactivate the instance then its buffers are destroyed.
But after destroying all buffers, it doesn't re-create after
re-activated the instances.

This patch fix it.

Change-Id: I36367a8c6ce11ad597b899d0841b73e52f2385a2