Remove verbose log message
Add SHM type FB management code (slave is not preapred yet)
Replace libconnector with libcore-com
Fix some bug of instance management code.
Try to send a packet again to a slave if it fails but the slave is not faulted.
Fix the com-core handle manage code (Reset it if fails to communicate w/t other)

Change-Id: I43620520329f158aab620e5fe071fe52d9812ae3