Acquire buffer/Release buffer implemented.
Slave can acquire the buffer of livebox and PD only if its type is
buffer_livebox, buffer_pd or the slave couldn't get any buffer of livebox or PD.

To support this feature,
livebox should use the "buffer_livebox, buffer_pd" option in its conf
file.

then the master will prepare the buffer and share it with the slave.

Change-Id: Id08a86184f668af9de3842cd8edce655f1df6099