Find proper slave for specified ABI.
And add missing file (buffer_handler.c)
Now the slave_activate function will launch the
proper slave for given ABI.

if the "abi" is "HTML", the master will launch the web slave.
or if the "abi" is "OSP", the master will launch the osp slave.
if the "abi" is "c" or "cpp" the master will launch the native slave.

Change-Id: If87a905d1de71f45e4f8de8a78e320abdb358a4c