If the master wants to terminate a slave,
Make sure that the slave is not faulted even if the master gets the
disconnected event.

When you call the slave_deactivate function, it will terminate the
slave.

Then the slave will be disconnected, it will generate the "disconnect"
event, the master will do the rest of a resource cleaning procedure.

Previously, the master tries to cleans all resource right after
call the aul_terminate_pid function.

The the master will receive the disconnected event again.
And the wrost cases, the slave can tries to do some remains job
for cleaning resources when it gets the terminates signal.

But the master already cleans all information of slave,
So the master cannot recognize the slave's last requests.

This patch will fix this weird termination scenario for the slave.

Change-Id: Iecc655a2cafd3fc690f95fc1f019ffff230367d3