Fix the implicit memory corruping bug.
Before this patch,
When a user tries to delete a livebox instance,
It couldn't care the slave object of a package information correctly.

slave_unload_instance can destroy the slave object.
But package_del_instance tries to access the slave object.

So I change the sequece of these two function calls.
To keep the slave object in safely.

Change-Id: I69d929780a6715af4989b775842bc21232674db7