Revise the client connection management code.
When the master creating a new data for clients,
it already knows the handle for communicating with the client.

Set the handle when it creates a new client data.

And call the created event callback after feed-back the acquire request.

Use the timer(DELAY_TIME) to call the created event callbacks.

Change-Id: Ief26b2efbb172de04e327e8d35c0e932d8fb5ba7