Release the resources while terminating the daemon.
RPC of a slave/client is now cleared by the destroy function.
Previously, To release the rpc resource, the delete callback is used,
and it is called from the rpc context.

Now this patch will change the sequence of it.
Creating RPC will be processed in the create function.
Destroying RPC will be processed in destroy function.

This is more intuitive for me.
But the RPC and each Data has to be paired all its life time.

Change-Id: I20da5db6836999ac300eee7a3a58a6df6b36719c