Update instance visibility handling code.
Only the LB_HIDE will block the slave's update event.

LB_PAUSE_WITH_HIDE will not block the slave's update event.
So if a livebox updates its contents, the updated event will be sent
to the viewer.
Because it needs to keep the latest contents
Update timer will be freezed. there should be no update contents.

Change-Id: Ibfb4a9c3f8c287eeff7a885560eb677d4b887b2c