Fix the reactivate + resize bug.
In case of the resize the livebox,
If the slave is terminated (secured),
the master tries to reactivate the slave and its liveboxes.
And it requests to resize the livebox.

After that the master send resize request again,
to recover the original size.

So the livebox cannot change the size correctly,.
Remove the resize function from the instance reactivated callback.

Change-Id: I1b85a2d6dc3d8ce64c14783ab23ff556e7712b4c