Fix the inactivated slave handling code.
There is a leak of resource in a very specific case (definitly rare case).
All request package will be pended at slave rpc->pending_list.

If the slave is not activated (didn't send hello packet),
The provider will terminate slave,
Then the slave rpc's deactivated callback will be invoked.

But the in there, slave rpc didn't clear the rpc->pending_list.
which packets are requested while activating the slave.

In a very specific case that the slave didn't send the hello message,
the rpc->pending_list can holds packets but previously, that list is
not cleared.

So this patch will clear it.

Furthermore,
If a slave is deactivated, invoke the ret_cb all of command packets.
To clear all registered callback data correctly.

Change-Id: I64e7275a3a36efaf781b3137d381ac48ff4bc401