Send PD close event to the slave if it fails.
If the provider(master) failes to load the EDJE of PD,
send the failed event to the slave not only to the client.

Before apply this patch, the slave couldn't know what happens
when the master couldn't open a PD (failed to load EDJE),
So the slave always pending all update reuqests.

Change-Id: I4767628bfa69cf2d082f69a0f2d62c67908af0e1