Keep the desc block even if the buffer of lb is not ready.
Keep the desc blocks even if the buffer of livebox is not created.
The provider will send the created event to the slave.

Then the slave will start to generate the desc file. (script data)

But the provider cannot apply desc to its buffer.
Because there is no created object.

To resolve this issue,
This patch will create a cache.

It will be used for keeping the blocks which are parsed before creating a buffer of livebox (or PD).

Change-Id: I27ee6fda35f93622eef08216da905b9e1ed0b5c7