Fix the bug of upgrading package.
While upgrading packages, the instance's update period doesn't updated correctly.
This patch will apply the new update period.

But if the instance's update period is changed manually, (that is not same with in the XML),
the master will not touch it.

Only if the period is same with orignal value in the old XML, will be updated to the new value.

Change-Id: I4e8e9f8a104a3010b6261ab77770960463051c84