Master should check the package name of slave.
In case of "app", if it uses "REPLACE_TAG", the package name of slave process
can be changed. in that case, we have to manage it correctly.

previously, master only checks "abi" for finding available service provider.
So if the provider uses "app" tag, every single instances requests are passed to a first "app" type provider.
But its package name is not same.

This patch will check the package name too.
So even though the ABI is same, if its package name is different, the master will route the request packets to
proper service provider.

Change-Id: If4a3159389ff1403236bd5dd6abed1277d9a1f06