Fix PLM P140811-03536
Even though the event_deactivate is called,
The event queue is not cleared yet.

Before the event queue is empty,
if the instance is deleted,
The event handler is able to try access the deleted instance.

To prevent this.
after activate the event thread,
do not remove the instance deletion callback.
It will be deleted only if the instance is destroyed.

After apply this patch we can reset instance pointer in the event queue.
When the instance gets destroy event, its delete callback will be called.
And then the delete-callback will reset all instance pointer in the event queue.

Change-Id: I43cad97be82b36e3b8bd96ffa1d49d09ca84d14f