Verify the slave's request.
In case of standalone type provider,
Master should validate the requestor's package name and requested dbox id.

The standalone provider can only request its dynamicbox's resource.
Or the master should reject it.

But this can increase the compliexity and decrease the performance.
So we should optimize the cost of verifying the requestor's identification logic.

TODO: Optimize the provider app verification logic

[model] Redwood,Kiran,B3(Wearable)
[binary_type] AP
[customer] Docomo/Orange/ATT/Open
[issue#] N/A
[problem]
[cause]
[solution]
[team] HomeTF
[request]
[horizontal_expansion]

Change-Id: Ie8e5ff7bc131869cdcbb1792ac7437779df7da39