Fix the duplicated active/relaunch timer registration issue.
When the slave is not activated (waiting hello message or relaunching it)
If a user adds widget instance again,

the slave "pause" / "resume" state change request can be sent.

At that time, the slave state is changed to REQUEST_PAUSE or REQUEST_RESUME.
but slave activate function only check the REQUEST_LAUNCH state.
and its pid is not speicified.

Because of this weird cases,
the slave gets two or more activate/reluanch timer.

it creates dangled timer pointer.

After few secs later, even if the slave is faulted. so the object is deleted.
those dangled timer object fired, and tries to access the deleted slave object.

finally, the master gets crashed.

This patch will fix it.

Change-Id: Ie418a874def0711ad2a8d3d6a01e9409e23abd8d